Authorization for delete comment

+++ pages/api/comment/delete/[id].js
@@ -1,8 +1,14 @@
 import prisma from "../../../../lib/prisma.ts"
+import { checkAuth } from "../../../../lib/auth";
 
 export default async (req, res) => {
+  const isAuth = await checkAuth(req);
+  if(!isAuth){
+    res.status(403).json({ err: "Forbidden" });
+    return;
+  }
   const {id} = req.query
-  //récupérer l'ID
+  //get ID
   try {
     const deleteComment = await prisma.comment.delete({
       where: {
